requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
huggingface-hub<0.25.0